Exception handling for problems with importer.

+++ src/main/java/com/zuehlke/pgadmissions/dto/ApplicationRatingDTO.java
@@ -1,27 +1,25 @@
 package com.zuehlke.pgadmissions.dto;
 
-import java.math.BigDecimal;
-
 public class ApplicationRatingDTO {
 
-    Integer ratingCount;
+    Long ratingCount;
     
-    BigDecimal ratingAverage;
+    Double ratingAverage;
 
-    public final Integer getRatingCount() {
+    public final Long getRatingCount() {
         return ratingCount;
     }
 
-    public final void setRatingCount(Integer ratingCount) {
+    public final void setRatingCount(Long ratingCount) {
         this.ratingCount = ratingCount;
     }
 
-    public final BigDecimal getRatingAverage() {
+    public final Double getRatingAverage() {
         return ratingAverage;
     }
 
-    public final void setRatingAverage(BigDecimal ratingAverage) {
+    public final void setRatingAverage(Double ratingAverage) {
         this.ratingAverage = ratingAverage;
     }
-    
+
 }
